refactor: 보드 초기화와 복원을 인터페이스로 추상화
